ADF-1754 Fix incompatibility with the twenty twenty-two theme
The theme uses webpacker to pack and deploy scripts which
caused the `document` passed into our script to not be the
DOM document but some kind of webpacker object.

The fix is to include the addressfinder javascript in the footer
instead of the header which causes the script to be loaded
the old and "boring" way.