ADF-1986 Fix NZ regions not populating
The element we were checking for the NZ region mapping
does not exist at the point in time of the check so we
always defaulted to 2-letter regions.

I have flipped the conditions around so that if the
checked element is not there yet, or does not contain
options with 2-letter values we return the 3-letter
mapping.