ADF-2504 Workaround for the select2 plugin issue
Two customer websites have reported issues with our
plugin which are caused by the select2 jquery plugin
used to fancy up various select boxes.

The HTML comes in with #shipping_state/billing_state
inputs as `type=text` and then it creates the select
for regions/states once it loads the data from an ajax
call.

Our plugin expects the input to be always a select
so it blows up. It is possible the issue will be fixed
with the commit properly if the mutationEventHandler
reloads the config few moments after.